% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{browseKosis}
\alias{browseKosis}
\title{Browse KOSIS webpage to find the url}
\usage{
browseKosis()
}
\value{
No return value, browse KOSIS webpage to find the url
}
\description{
Browse KOSIS webpage to find the url
}
\details{
\preformatted{
## Example
browseKosis()
}
}
