% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{castItem}
\alias{castItem}
\title{Cast Item variable from the stat data downloaded from getStatData function}
\usage{
castItem(statData, itemVar = c("ITM_NM", "ITM_ID", "ITM_NM_ENG"))
}
\arguments{
\item{statData}{A data.frame downloaded from getStatData function}

\item{itemVar}{A string specifying item variable}
}
\value{
A data.frame object containing queried information
}
\description{
Cast Item variable from the stat data downloaded from getStatData function
}
\details{
\preformatted{
## Example
statData <- getStatData(orgId = "101", tblId = "DT_1B41", objL1 = "ALL")
castItem(statData = statData, itemVar = "ITM_NM")
}
}
