% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStatData.R
\name{getStatDataFromURL}
\alias{getStatDataFromURL}
\title{KOSIS Statistic Data from URL}
\usage{
getStatDataFromURL(url)
}
\arguments{
\item{url}{A string specifying the KOSIS URL}
}
\value{
A data.frame object containing queried information
}
\description{
KOSIS Statistic Data from URL
}
\details{
\preformatted{
## Example
url <- "https://kosis.kr/openapi/Param/statisticsParameterData.do?
method=getList&apiKey=&
itmId=T6+T16+T26+T5+T15+T25+T2+T12+T22+T3+T13+T23+T4+T14+T24+T1+T11+T21+&
objL1=ALL&objL2=&objL3=&objL4=&objL5=&objL6=&objL7=&objL8=&
format=json&jsonVD=Y&
prdSe=Y&newEstPrdCnt=3&prdInterval=1&
orgId=101&tblId=DT_1B41"
getStatDataFromURL(url = url)
}
}
