% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStatExpl.R
\name{getStatExpl}
\alias{getStatExpl}
\title{KOSIS Statistic Explanation}
\usage{
getStatExpl(orgId, tblId, statId, metaItm = "ALL")
}
\arguments{
\item{orgId}{A string specifying the organization id}

\item{tblId}{A string specifying the table id}

\item{statId}{A string specifying the stat id}

\item{metaItm}{A string specifying the meta item}
}
\value{
A data.frame object containing queried information
}
\description{
KOSIS Statistic Explanation
}
\details{
\preformatted{
## Example
getStatExpl(orgId = "101", tblId = "DT_1L9V054")
getStatExpl(statId = "1962009")
}
}
