% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStatList.R
\name{getStatList}
\alias{getStatList}
\title{KOSIS Statistic List}
\usage{
getStatList(vwCd = "MT_ZTITLE", parentListId)
}
\arguments{
\item{vwCd}{A string specifying the view code}

\item{parentListId}{A string specifying the parent list id}
}
\value{
A data.frame object containing queried information
}
\description{
KOSIS Statistic List
}
\details{
\preformatted{
## Example
getStatList(vwCd = "MT_ZTITLE", parentListId = "A")
}
}
