% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStatMeta.R
\name{getStatMeta}
\alias{getStatMeta}
\title{KOSIS Statistic Meta}
\usage{
getStatMeta(
  orgId,
  tblId,
  type = c("TBL", "ORG", "PRD", "ITM", "CMMT", "UNIT", "SOURCE", "WGT", "NCD")
)
}
\arguments{
\item{orgId}{A string specifying the organization id}

\item{tblId}{A string specifying the table id}

\item{type}{A string specifying the meta type}
}
\value{
A data.frame object containing queried information
}
\description{
KOSIS Statistic Meta
}
\details{
\preformatted{
## Example
getStatMeta(orgId = "101", tblId = "DT_1IN0001")
getStatMeta(orgId = "101", tblId = "DT_1IN0001", type = "ORG")
}
}
