% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStatSearch.R
\name{getStatSearch}
\alias{getStatSearch}
\title{KOSIS Statistic Search}
\usage{
getStatSearch(
  searchNm,
  startCount = 1,
  resultCount = 20,
  sort = c("RANK", "DATE")
)
}
\arguments{
\item{searchNm}{A string specifying a name you search}

\item{startCount}{A nuemeric specifying a page number}

\item{resultCount}{A numeric specifying counts of result
(resultCount = 20, startCount = 2 : 21~40 Result)}

\item{sort}{A string specifying the sorting (RANK: accuracy, DATE: newest)}
}
\value{
A data.frame object containing queried information
}
\description{
KOSIS Statistic Search
}
\details{
\preformatted{
## Example
getStatSearch(searchNm = "CPI")
}
}
