% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kosis-package.R
\docType{package}
\name{kosis-package}
\alias{kosis}
\alias{kosis-package}
\title{kosis: Korean Statistical Information Service (KOSIS)}
\description{
API wrapper to download statistical information from the Korean Statistical Information Service (KOSIS) \url{https://kosis.kr/openapi/index/index.jsp}.
}
\details{
To use this package, you will first need to get your API key from the
website \url{https://kosis.kr/openapi/index/index.jsp}. Once you have your
key, you can save it as an environment variable for the current session
using the \code{\link{kosis.setKey}} function. Alternatively, you can set it
permanently by adding the following line to your .Renviron file:

KOSIS_API_KEY = PASTE YOUR API KEY

Any functions that require your API key try to retrieve it via
\code{Sys.getenv("KOSIS_API_KEY")} (unless API key is explicitly specified as
a function argument).
}
\author{
\strong{Maintainer}: Seokhoon Joo \email{seokhoonj@gmail.com}

}
