% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setKey.R
\name{kosis.setKey}
\alias{kosis.setKey}
\alias{kosis.printKey}
\title{Set KOSIS API Key}
\usage{
kosis.setKey(apiKey)

kosis.printKey()
}
\arguments{
\item{apiKey}{A string specifying KOSIS API key}
}
\value{
No return value, called to set api key
}
\description{
Save KOSIS API key for the current session. To set it permanently, please add
the following line to your .Renvrion file:

KOSIS_API_KEY = "YOUR API KEY"
}
\examples{

## Set API Key for the current session
\donttest{kosis.setKey("your_api_key")}

## Check API key
kosis.printKey()

}
