% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{removeUniqueCols}
\alias{removeUniqueCols}
\title{Remove columns that consist of unique values}
\usage{
removeUniqueCols(statData)
}
\arguments{
\item{statData}{A data.frame downloaded from getStatData function}
}
\value{
A data.frame object with columns that consist of non-unique values
}
\description{
Remove columns that consist of unique values
}
\details{
\preformatted{
## Example
df <- removeUniqueCols(statData = statData)
}
}
