% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStatList.R
\name{getStatList}
\alias{getStatList}
\title{KOSIS Statistic List}
\usage{
getStatList(
  vwCd = c("MT_ZTITLE", "MT_OTITLE", "MT_GTITLE01", "MT_GTITLE02", "MT_CHOSUN_TITLE",
    "MT_HANKUK_TITLE", "MT_STOP_TITLE", "MT_RTITLE", "MT_BUKHAN", "MT_TM1_TITLE",
    "MT_TM2_TITLE", "MT_ETITLE"),
  parentListId = ""
)
}
\arguments{
\item{vwCd}{A string specifying the view code}

\item{parentListId}{A string specifying the parent list id}
}
\value{
A data.frame object containing queried information
}
\description{
KOSIS Statistic List
}
\details{
\preformatted{
## Example
getStatList(vwCd = "MT_ZTITLE", parentListId = "")
getStatList(vwCd = "MT_ZTITLE", parentListId = "F")
getStatList(vwCd = "MT_ZTITLE", parentListId = "F_29")
}
}
